refactor: take attendance hoook
