--- apps/web/src/hooks/queries/useAttendances/queryKey.ts
@@ -1,2 +1,0 @@
-export const useQueryKey = (orgId: string) =>
-  ["orgs", orgId, 'attendances'] as const;
